server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
tensorflow==2.18.0
tensorflow-hub==0.16.1
numpy
mediapipe==0.10.21
opencv-python==4.8.1.78 